requirements.txt
numpy>=1.21.5
pandas>=1.1.5
torch>=1.7.0
torchvision>=0.11.1
monai>=0.8.1
pillow
scipy
yaml
json